Struct google_domains1_beta1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudDomains hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_domains1_beta1 as domains1_beta1;
 
use std::default::Default;
use oauth2;
use domains1_beta1::CloudDomains;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudDomains::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_get(...)`, `locations_list(...)`, `locations_operations_get(...)`, `locations_operations_list(...)`, `locations_registrations_configure_contact_settings(...)`, `locations_registrations_configure_dns_settings(...)`, `locations_registrations_configure_management_settings(...)`, `locations_registrations_delete(...)`, `locations_registrations_export(...)`, `locations_registrations_get(...)`, `locations_registrations_get_iam_policy(...)`, `locations_registrations_list(...)`, `locations_registrations_patch(...)`, `locations_registrations_register(...)`, `locations_registrations_reset_authorization_code(...)`, `locations_registrations_retrieve_authorization_code(...)`, `locations_registrations_retrieve_register_parameters(...)`, `locations_registrations_search_domains(...)`, `locations_registrations_set_iam_policy(...)` and `locations_registrations_test_iam_permissions(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_registrations_configure_contact_settings(
&self,
request: ConfigureContactSettingsRequest,
registration: &str
) -> ProjectLocationRegistrationConfigureContactSettingCall<'a, C>

Create a builder to help you perform the following task:

Updates a Registration's contact settings. Some changes require confirmation by the domain's registrant contact .

Arguments

request - No description provided.

registration - Required. The name of the Registration whose contact settings are being updated, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_configure_dns_settings(
&self,
request: ConfigureDnsSettingsRequest,
registration: &str
) -> ProjectLocationRegistrationConfigureDnsSettingCall<'a, C>

Create a builder to help you perform the following task:

Updates a Registration's DNS settings.

Arguments

request - No description provided.

registration - Required. The name of the Registration whose DNS settings are being updated, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_configure_management_settings(
&self,
request: ConfigureManagementSettingsRequest,
registration: &str
) -> ProjectLocationRegistrationConfigureManagementSettingCall<'a, C>

Create a builder to help you perform the following task:

Updates a Registration's management settings.

Arguments

request - No description provided.

registration - Required. The name of the Registration whose management settings are being updated, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_delete(
&self,
name: &str
) -> ProjectLocationRegistrationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a Registration resource. This method only works on resources in one of the following states: * state is EXPORTED with expire_time in the past * state is REGISTRATION_FAILED

Arguments

name - Required. The name of the Registration to delete, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_export(
&self,
request: ExportRegistrationRequest,
name: &str
) -> ProjectLocationRegistrationExportCall<'a, C>

Create a builder to help you perform the following task:

Exports a Registration that you no longer want to use with Cloud Domains. You can continue to use the domain in Google Domains until it expires. If the export is successful: * The resource's state becomes EXPORTED, meaning that it is no longer managed by Cloud Domains * Because individual users can own domains in Google Domains, the calling user becomes the domain's sole owner. Permissions for the domain are subsequently managed in Google Domains. * Without further action, the domain does not renew automatically. The new owner can set up billing in Google Domains to renew the domain if needed.

Arguments

request - No description provided.

name - Required. The name of the Registration to export, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_get(
&self,
name: &str
) -> ProjectLocationRegistrationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the details of a Registration resource.

Arguments

name - Required. The name of the Registration to get, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationRegistrationGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_registrations_list(
&self,
parent: &str
) -> ProjectLocationRegistrationListCall<'a, C>

Create a builder to help you perform the following task:

Lists the Registration resources in a project.

Arguments

parent - Required. The project and location from which to list Registrations, specified in the format projects/*/locations/*.

pub fn locations_registrations_patch(
&self,
request: Registration,
name: &str
) -> ProjectLocationRegistrationPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates select fields of a Registration resource, notably labels. To update other fields, use the appropriate custom update method: * To update management settings, see ConfigureManagementSettings * To update DNS configuration, see ConfigureDnsSettings * To update contact information, see ConfigureContactSettings

Arguments

request - No description provided.

name - Output only. Name of the Registration resource, in the format projects/*/locations/*/registrations/.

pub fn locations_registrations_register(
&self,
request: RegisterDomainRequest,
parent: &str
) -> ProjectLocationRegistrationRegisterCall<'a, C>

Create a builder to help you perform the following task:

Registers a new domain name and creates a corresponding Registration resource. Call RetrieveRegisterParameters first to check availability of the domain name and determine parameters like price that are needed to build a call to this method. A successful call creates a Registration resource in state REGISTRATION_PENDING, which resolves to ACTIVE within 1-2 minutes, indicating that the domain was successfully registered. If the resource ends up in state REGISTRATION_FAILED, it indicates that the domain was not registered successfully, and you can safely delete the resource and retry registration.

Arguments

request - No description provided.

parent - Required. The parent resource of the Registration. Must be in the format projects/*/locations/*.

pub fn locations_registrations_reset_authorization_code(
&self,
request: ResetAuthorizationCodeRequest,
registration: &str
) -> ProjectLocationRegistrationResetAuthorizationCodeCall<'a, C>

Create a builder to help you perform the following task:

Resets the authorization code of the Registration to a new random string. You can call this method only after 60 days have elapsed since the initial domain registration.

Arguments

request - No description provided.

registration - Required. The name of the Registration whose authorization code is being reset, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_retrieve_authorization_code(
&self,
registration: &str
) -> ProjectLocationRegistrationRetrieveAuthorizationCodeCall<'a, C>

Create a builder to help you perform the following task:

Gets the authorization code of the Registration for the purpose of transferring the domain to another registrar. You can call this method only after 60 days have elapsed since the initial domain registration.

Arguments

registration - Required. The name of the Registration whose authorization code is being retrieved, in the format projects/*/locations/*/registrations/*.

pub fn locations_registrations_retrieve_register_parameters(
&self,
location: &str
) -> ProjectLocationRegistrationRetrieveRegisterParameterCall<'a, C>

Create a builder to help you perform the following task:

Gets parameters needed to register a new domain name, including price and up-to-date availability. Use the returned values to call RegisterDomain.

Arguments

location - Required. The location. Must be in the format projects/*/locations/*.

pub fn locations_registrations_search_domains(
&self,
location: &str
) -> ProjectLocationRegistrationSearchDomainCall<'a, C>

Create a builder to help you perform the following task:

Searches for available domain names similar to the provided query. Availability results from this method are approximate; call RetrieveRegisterParameters on a domain before registering to confirm availability.

Arguments

location - Required. The location. Must be in the format projects/*/locations/*.

pub fn locations_registrations_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationRegistrationSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_registrations_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationRegistrationTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.